[add] added a interface for reading from file, added dialog to select headers/keys
added a interface with static function that you can extends to choose the way you want to read froma file
Added a dialog that displays the keys/headers you give it/it has found from the Reader.
Then the reader read the file and return a list of Competitor objects that can be inserted in the tree
with their matching keys/headers values.